\name{opefimor-package}
\alias{opefimor-package}
\alias{opefimor}
\docType{package}
\title{
Option Pricing and Estimation of Financial Models in R
}
\description{
Companion package to the book Option Pricing and
		Estimation of Financial Models in R, Wiley, Chichester.
		ISBN: 978-0-470-74584-7.
}
\details{
This package contain all the code used in the book. The files are collected
by chapters 2 to 9 and the two appendixes (A and B).

The files are accessible from the folder \code{inst} of this package.
Please see the examples.
}
\author{Stefano M. Iacus}
\keyword{ package }
\examples{
\dontrun{
setwd(file.path(.libPaths()[1],"opefimor","book"))
dir()
}
}
